+++ storecassandra/metrics.go
@@ -5,12 +5,12 @@ import (
 	"tux21b.org/v1/gocql"
 )
 
-func (s *StoreCassandra) SaveMetric(metric string, value int) error {
+func (s *StoreCassandra) SaveMetric(metric string, value float64) error {
 	return s.session.Query(`INSERT INTO Metrics (key, value) VALUES (?, ?)`, metric, value).Exec()
 }
 
-func (s *StoreCassandra) GetMetric(metric string) (int, error) {
-	var value int
+func (s *StoreCassandra) GetMetric(metric string) (float64, error) {
+	var value float64
 	err := s.session.Query(`SELECT value FROM Metrics WHERE key=?`, metric).Scan(&value)
 
 	if err == gocql.ErrNotFound {
